% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-Nearest Neighbors Cross-Validation Error}
\usage{
my_knn_cv(train, cl, k_nn = 1, k_cv)
}
\arguments{
\item{train}{Input data frame to be used for training and test data.}

\item{cl}{Column of true class values to be used for making and testing
predictions. A string vector.}

\item{k_nn}{Integer representing number of neighbors used for predicting class
of test data.}

\item{k_cv}{Integer representing number of folds used for calculating
cross-validation error.}
}
\value{
A list with the following components:

\code{class}: Vector of predicted class for each observation of
\code{train} dataset.

\code{cv_err}: Integer for cross-validation misclassification error.
}
\description{
This function conducts k-nearest neighbors on a data frame and reports the
cross-validation misclassification test error.
}
\examples{
# Conduct k-nearest neighbor on penguins dataset

# Penguin training data with select covariates
penguin_train <- data.frame("bill_length_mm" = penguins$bill_length_mm,
                           "bill_depth_mm" = penguins$bill_depth_mm,
                           "flipper_length" = penguins$flipper_length_mm,
                           "body_mass_g" = penguins$body_mass_g)

# Penguin class information
penguin_cl <- penguins$species

# Test with k_nn = 1 nearest neighbor in penguins dataset
my_knn_cv(penguin_train, penguin_cl, 5)

# Test with k_nn = 5 nearest neighbors in penguins dataset
my_knn_cv(penguin_train, penguin_cl, 5, 5)

}
\keyword{prediction}
\keyword{statistical}
