% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object relating to covariates}

\item{data}{Input data frame.}
}
\value{
A table with the following column entries for each coefficient:

\code{estimate}: For each beta coefficient and the intercept.

\code{st_err}: Standard error on each coefficient.

\code{t_value}: A test statistic for each coefficient.

\code{p_value}: Probability of yielding a \code{t_value} this
extreme or greater.
}
\description{
This function produces a linear regression model, yielding a table with the
estimate, standard error, t-value, and p-value for each coefficient.
}
\examples{
my_lm(mpg ~ hp + wt, data = mtcars)

}
\keyword{inference}
\keyword{prediction}
\keyword{statistical}
