% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation Error}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k_cv}{Integer representing number of folds used for calculating
cross-validation error.}
}
\value{
\code{cv_err}: Integer for cross-validation error.
}
\description{
This function conducts a random forest analysis of the penguins dataset and
reports the cross-validation misclassification test error.
}
\examples{
# Test with k = 5 folds
my_rf_cv(5)

}
\keyword{prediction}
\keyword{statistical}
