% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{Student's T-test}
\usage{
my_t.test(x, mu = 0, alternative = "two.sided")
}
\arguments{
\item{x}{Numeric vector of data to be tested.}

\item{mu}{Numeric input indicating the mean under the null distribution.
Defaults to 0.}

\item{alternative}{String input indicating if test is "\code{two.sided}",
"\code{less}", or "\code{greater}". Defaults to "\code{two.sided}".}
}
\value{
A list with the following components:

\code{test_stat}: A test statistic

\code{df}: Degrees of freedom used in the test.

\code{alternative}: String indicating if test is "\code{two.sided}",
"\code{less}", or "\code{greater}".

\code{p_value}: Probability of yielding a \code{test_stat} this
extreme or greater.
}
\description{
This function performs a t-test on a numeric vector, yielding the test
statistic, degrees of freedom, type of test, and p-value.
}
\examples{
data_1 <- rnorm(10, 3, 2)
my_t.test(data_1)

data_2 <- c(0, 0, 0, 0, 50)
my_t.test(data_2, 1, "greater")

}
\keyword{inference}
\keyword{statistical}
