% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_fun.R
\name{sum_fun}
\alias{sum_fun}
\title{Perform addition}
\usage{
sum_fun(x, y)
}
\arguments{
\item{x}{A numeric variable}

\item{y}{A numeric variable}
}
\value{
A numeric variable
}
\description{
Perform addition
}
\examples{
x <- 3
y <- 4
sum_fun(x,y)
}
